+++ docs/src/changelog.md
@@ -7,6 +7,13 @@ CurrentModule = LibSSH
 This documents notable changes in LibSSH.jl. The format is based on [Keep a
 Changelog](https://keepachangelog.com).
 
+## Unreleased
+
+### Added
+
+- [`Demo.DemoServer`](@ref) now supports passing `allow_auth_none=true` to allow
+  easily setting up passwordless authentication ([#28]).
+
 ## [v0.6.1] - 2024-10-20
 
 ### Added